requirements.txt
matplotlib
pandas
numpy
requests        
pyyaml
networkx        
sympy           
tqdm
torch
torchaudio
Pillow
soundfile
pydub
chardet